// src/generated/models/request-model.ts
export const RequestStatusKeyToLabel = {
  'New': 'New',
  'Assigned': 'Assigned',
  'InProgress': 'In Progress',
  'Pending': 'Pending',
  'Resolved': 'Resolved',
  'Closed': 'Closed',
  'Cancelled': 'Cancelled'
} as const;
export type RequestStatusKey = keyof typeof RequestStatusKeyToLabel;

export interface Request {
// ...
  statusKey: RequestStatusKey;
// ...
}