// src/generated/models/status-transition-model.ts
import type { StatusDefinition } from './status-definition-model';
import type { LifecycleDefinition } from './lifecycle-definition-model';
import type { Workspace } from './workspace-model';

export const StatusTransitionStatusKeyToLabel = {
  'Active': 'Active',
  'Inactive': 'Inactive'
} as const;
export type StatusTransitionStatusKey = keyof typeof StatusTransitionStatusKeyToLabel;

export interface StatusTransition {
  /**
   * @displayName Status Transition ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Transition Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  transitionName: string;
  /**
   * @displayName Destination Status Definition ID
   * @validationRule Required for create/update operations
   */
  destinationStatusDefinitionID: Pick<StatusDefinition, 'id' | 'statusName'>;
  /**
   * @displayName Lifecycle Definition ID
   * @validationRule Required for create/update operations
   */
  lifecycleDefinitionID: Pick<LifecycleDefinition, 'id' | 'name1'>;
  /**
   * @displayName Sort Order
   * @validationRule Required for create/update operations
   */
  sortOrder: number;
  /**
   * @displayName Source Status Definition ID
   * @validationRule Required for create/update operations
   */
  sourceStatusDefinitionID: Pick<StatusDefinition, 'id' | 'statusName'>;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: StatusTransitionStatusKey;
  /**
   * @displayName Transition Code
   * @validationRule Required for create/update operations
   */
  transitionCode: string;
  /**
   * @displayName Workspace ID
   * @validationRule Required for create/update operations
   */
  workspaceID: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _StatusTransition = 'StatusTransition' as const;