// apps/enterprise-service-management/src/generated/models/lifecycle-transition-required-field-model.ts
import type { StatusTransition } from './status-transition-model';
import type { Workspace } from './workspace-model';

export const LifecycleTransitionRequiredFieldStatusKeyToLabel = {
  'Active': 'Active',
  'Inactive': 'Inactive'
} as const;
export type LifecycleTransitionRequiredFieldStatusKey = keyof typeof LifecycleTransitionRequiredFieldStatusKeyToLabel;

export interface LifecycleTransitionRequiredField {
  /**
   * @displayName Lifecycle Transition Required Field ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Field Label
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  fieldLabel: string;
  /**
   * @displayName Field Name
   * @validationRule Required for create/update operations
   */
  fieldName: string;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: LifecycleTransitionRequiredFieldStatusKey;
  /**
   * @displayName Transition
   * @validationRule Required for create/update operations
   */
  transition: Pick<StatusTransition, 'id' | 'transitionName'>;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _LifecycleTransitionRequiredField = 'LifecycleTransitionRequiredField' as const;