// src/generated/models/request-task-model.ts
import type { Person } from './person-model';
import type { Request } from './request-model';
import type { Workspace } from './workspace-model';

export const RequestTaskStatusIdKeyToLabel = {
  'NotStarted': 'Not Started',
  'InProgress': 'In Progress',
  'Completed': 'Completed',
  'Blocked': 'Blocked',
  'Cancelled': 'Cancelled'
} as const;
export type RequestTaskStatusIdKey = keyof typeof RequestTaskStatusIdKeyToLabel;

export interface RequestTask {
  /**
   * @displayName Request Task ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Title
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  title: string;
  /**
   * @displayName AssignedPersonId
   */
  assignedPersonId?: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName CompletedAt
   */
  completedAt?: string;
  /**
   * @displayName DueDate
   */
  dueDate?: string;
  /**
   * @displayName RequestId
   * @validationRule Required for create/update operations
   */
  requestId: Pick<Request, 'id' | 'requestNumber'>;
  /**
   * @displayName StatusId
   * @validationRule Required for create/update operations
   */
  statusIdKey: RequestTaskStatusIdKey;
  /**
   * @displayName WorkspaceId
   * @validationRule Required for create/update operations
   */
  workspaceId: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _RequestTask = 'RequestTask' as const;