// apps/enterprise-service-management/src/generated/models/request-approval-model.ts
import type { ApprovalDefinition } from './approval-definition-model';
import type { Person } from './person-model';
import type { Role } from './role-model';
import type { Request } from './request-model';
import type { Workspace } from './workspace-model';

export const RequestApprovalApproverTypeKeyToLabel = {
  'Manager': 'Manager',
  'Person': 'Person',
  'Role': 'Role',
  'Group': 'Group'
} as const;
export type RequestApprovalApproverTypeKey = keyof typeof RequestApprovalApproverTypeKeyToLabel;

export const RequestApprovalStatusKeyToLabel = {
  'Waiting': 'Waiting',
  'Pending': 'Pending',
  'Approved': 'Approved',
  'Rejected': 'Rejected',
  'Expired': 'Expired',
  'Escalated': 'Escalated',
  'Cancelled': 'Cancelled'
} as const;
export type RequestApprovalStatusKey = keyof typeof RequestApprovalStatusKeyToLabel;

export interface RequestApproval {
  /**
   * @displayName Request Approval ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName ApprovalName
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  approvalName: string;
  /**
   * @displayName Approval Definition
   */
  approvalDefinition?: Pick<ApprovalDefinition, 'id' | 'name1'>;
  /**
   * @displayName Approver Group Code
   */
  approverGroupCode?: string;
  /**
   * @displayName ApproverPersonId
   * @validationRule Required for create/update operations
   */
  approverPersonId: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Approver Role
   */
  approverRole?: Pick<Role, 'id' | 'roleName'>;
  /**
   * @displayName Approver Type
   * @validationRule Required for create/update operations
   */
  approverTypeKey: RequestApprovalApproverTypeKey;
  /**
   * @displayName Comments
   * @validationRule Required for create/update operations
   */
  comments: string;
  /**
   * @displayName CreatedAt
   * @validationRule Required for create/update operations
   */
  createdAt: string;
  /**
   * @displayName DecidedAt
   * @validationRule Required for create/update operations
   */
  decidedAt: string;
  /**
   * @displayName Delegated From Person
   */
  delegatedFromPerson?: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName RequestId
   * @validationRule Required for create/update operations
   */
  requestId: Pick<Request, 'id' | 'requestNumber'>;
  /**
   * @displayName Stage Number
   * @validationRule Required for create/update operations
   */
  stageNumber: number;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: RequestApprovalStatusKey;
  /**
   * @displayName WorkspaceId
   * @validationRule Required for create/update operations
   */
  workspaceId: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _RequestApproval = 'RequestApproval' as const;