// apps/enterprise-service-management/src/generated/models/request-resolution-model.ts
import type { Request } from './request-model';
import type { Person } from './person-model';
import type { Workspace } from './workspace-model';

export const RequestResolutionResolutionCodeKeyToLabel = {
  'Fulfilled': 'Fulfilled',
  'Fixed': 'Fixed',
  'Workaround': 'Workaround',
  'InformationProvided': 'Information Provided',
  'Duplicate': 'Duplicate',
  'Cancelled': 'Cancelled'
} as const;
export type RequestResolutionResolutionCodeKey = keyof typeof RequestResolutionResolutionCodeKeyToLabel;

export interface RequestResolution {
  /**
   * @displayName Request Resolution ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Request Resolution Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  requestResolutionName: string;
  /**
   * @displayName Request
   * @validationRule Required for create/update operations
   */
  request: Pick<Request, 'id' | 'requestNumber'>;
  /**
   * @displayName Resolution Code
   * @validationRule Required for create/update operations
   */
  resolutionCodeKey: RequestResolutionResolutionCodeKey;
  /**
   * @displayName Resolved At
   * @validationRule Required for create/update operations
   */
  resolvedAt: string;
  /**
   * @displayName Resolved By
   * @validationRule Required for create/update operations
   */
  resolvedBy: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Summary
   * @validationRule Required for create/update operations
   */
  summary: string;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _RequestResolution = 'RequestResolution' as const;